Revert "[Improve][API] Unify type system api(data & type)"
This reverts commit 800817fe3aadd82ffcd6d814705a27cb686df835.

+++ seatunnel-connectors-v2/connector-dolphindb-wt/src/main/java/org/apache/seatunnel/connectors/dolphindb/catalog/DolphinDBCatalog.java
@@ -167,7 +167,7 @@ public class DolphinDBCatalog implements Catalog {
                                 columnName,
                                 dolphinDBDataTypeConvertor.toSeaTunnelType(
                                         columnName, dataType.name()),
-                                (Long) null,
+                                null,
                                 true,
                                 null,
                                 null);
